move sorting functions to resource filtering modules
